Add trial number to MLflowCallback log_params
In mlflow we have the capability to create charts based on hpo
parameters and metrics, but we are missing the trial number as parameter
to for instance create chart of metrics over trials.
With this change we are logging the params just as before but we also
add the trial number as parameter.